requirements.txt
fastapi
uvicorn
python-multipart
jinja2
PyMuPDF
google-generativeai
python-dotenv
